requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

openai>=1.0.0
pyodbc
pandas
azure-functions
azure-storage-blob
pyarrow
streamlit
requests
openpyxl
